file_metadata_extractor/requirements.txt
json5~=0.12.1
langchain-community~=0.4.1
langchain-core~=1.0.4
langchain~=1.0.5
neo4j~=6.0.3
nltk~=3.9.2
numpy~=2.3.4
openai~=1.58.1
pydantic~=2.10.6
pypdf~=6.2.0
PyYAML~=6.0.3
RapidFuzz~=3.14.3
sentence-transformers~=5.1.2
structured-logprobs~=0.1.5
torch~=2.9.0
transformers~=4.57.1
fastapi
uvicorn
python-multipart